requirements.txt
redis
pymongo
pika
python-decouple
